+++ common/src/main/resources/assets/head_in_the_clouds/icon.png
