+++ fabric/src/main/resources/fabric.mod.json
@@ -1,0 +1,24 @@
+{
+  "schemaVersion": 1,
+  "id": "${mod_id}",
+  "version": "${version}",
+  "name": "${display_name}",
+  "description": "${description}",
+  "authors": [
+    "DaFuqs"
+  ],
+  "contact": {
+    "sources": "https://github.com/DaFuqs/HeadInTheClouds",
+    "issues": "https://github.com/DaFuqs/HeadInTheClouds/issues"
+  },
+  "license": "GPL-3.0-or-later",
+  "icon": "assets/${mod_id}/icon.png",
+  "environment": "client",
+  "mixins": [
+    "${mod_id}.mixins.json"
+  ],
+  "depends": {
+    "minecraft": ">=1.21.6",
+    "fabricloader": ">=0.15"
+  }
+}
