+++ gradle.properties
@@ -2,17 +2,18 @@
 org.gradle.jvmargs=-Xmx1G
 org.gradle.parallel=true
 
-# Fabric Properties
-# check these on https://modmuss50.me/fabric.html
+# Minecraft Properties
+# check these on https://fabricmc.net/develop
 minecraft_version=1.21.8
 yarn_mappings=1.21.8+build.1
-loader_version=0.17.2
-loom_version=1.11-SNAPSHOT
+fabric_loader_version=0.17.2
+neoforge_version=21.8.47
+yarn_mappings_patch_neoforge_version=1.21+build.4
 
 # Mod Properties
-mod_version=2.0.1+1.21.8
-maven_group=de.dafuqs
-archives_base_name=head_in_the_clouds
-
-# Dependencies
-fabric_api_version=0.131.0+1.21.8
+mod_version=2.1+1.21.8
+maven_group=de.dafuqs.head_in_the_clouds
+mod_id=head_in_the_clouds
+display_name=Head in the Clouds
+mod_description=Make it not rain over the cloud level
+enabled_platforms=fabric,neoforge
