+++ neoforge/src/main/resources/META-INF/neoforge.mods.toml
@@ -1,0 +1,31 @@
+modLoader = "lowcodefml"
+loaderVersion = "[1.0,)"
+issueTrackerURL = "https://github.com/DaFuqs/HeadInTheClouds/issues"
+license = "GPL-3.0-or-later"
+
+[[mods]]
+modId = "${mod_id}"
+version = "${version}"
+displayName = "${display_name}"
+authors = "DaFuqs"
+description = '''
+${description}
+'''
+logoFile = "assets/${mod_id}/icon.png"
+
+[[dependencies.${mod_id}]]
+modId = "neoforge"
+type = "required"
+versionRange = "[20,)"
+ordering = "NONE"
+side = "BOTH"
+
+[[dependencies.${mod_id}]]
+modId = "minecraft"
+type = "required"
+versionRange = "[1.21.6,)"
+ordering = "NONE"
+side = "BOTH"
+
+[[mixins]]
+config = "${mod_id}.mixins.json"
