--- src/main/resources/fabric.mod.json
@@ -1,30 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "head_in_the_clouds",
-  "version": "${version}",
-  "name": "Head In The Clouds",
-  "description": "Make it not rain over the cloud level",
-  "authors": [
-    "DaFuqs"
-  ],
-  "contact": {
-    "website": "https://github.com/DaFuqs/HeadInTheClouds",
-    "repo": "https://github.com/DaFuqs/HeadInTheClouds"
-  },
-  "license": "LGPLv3",
-  "icon": "assets/head_in_the_clouds/icon.png",
-  "environment": "*",
-  "entrypoints": {
-    "main": [
-      "de.dafuqs.head_in_the_clouds.HeadInTheClouds"
-    ]
-  },
-  "mixins": [
-    "head_in_the_clouds.mixins.json"
-  ],
-  "depends": {
-    "minecraft": ">=${minecraft_version}",
-    "fabricloader": ">=${loader_version}",
-    "java": ">=21"
-  }
-}
